Restructure rerun example into bridge and script
